+++ lib/multiutillib.dart
@@ -1,6 +1,6 @@
 library multiutillib;
 
-export 'constants.dart';
+export 'utils/constants.dart';
 
 //#region utils
 export 'utils/utils.dart';