+++ lib/widgets/connectivity_layout.dart
@@ -2,7 +2,7 @@
 
 import 'package:flutter/material.dart';
 
-import 'package:multiutillib/constants.dart';
+import 'package:multiutillib/utils/constants.dart';
 import 'package:connectivity_wrapper/connectivity_wrapper.dart';
 
 /// This widget is used to show messages for network operations